Rename upgrade slots helper and fix upgrades available to buy function
